Changed int to string

+++ src/main/java/com/joinef/eftrains/service/Edge.java
@@ -7,9 +7,9 @@ public class Edge {
 
     public final Vertex target;
 
-    public final double weight;
+    public final float weight;
 
-    public Edge(Vertex argTarget, double argWeight) {
+    public Edge(Vertex argTarget, float argWeight) {
         target = argTarget;
         weight = argWeight;
     }
